All Tasks Completed, Code cleanup needed and GUI fixes

+++ src/BackEnd/HTMLParser.java
@@ -9,6 +9,7 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.PrintStream;
+import java.sql.SQLOutput;
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.Map.Entry;
@@ -18,37 +19,40 @@ import GUI.UI;
 import static java.lang.Thread.sleep;
 
 public class HTMLParser {
-    static File ctrlFile = new File("control.txt");
-    static Scanner sc;
+    private static File ctrlFile = new File( "control.txt" );
+    private static Scanner scanControlFile;
+    public static String text;
+    public static boolean customLinkFlag;
 
     static {
         try {
-            sc = new Scanner(ctrlFile);
+            scanControlFile = new Scanner( ctrlFile );
         } catch (FileNotFoundException e) {
             e.printStackTrace();
         }
     }
 
-    static int amountOfLinks;
-
 
     public static void clearFiles() {
-        File folder = new File("/");
-        File fList[] = folder.listFiles();
+        String workingDir = System.getProperty( "user.dir" );
+        File folder = new File( workingDir );
+        File[] fList = folder.listFiles();
 
-        for (File file : fList) {
-            if (file.getName().endsWith(".txt")) {
-                file.delete();
-            } else if (file.getName().startsWith("control.txt")) {
-            }
-//TODO: FIX THIS
+        if (fList == null) {
+            throw new AssertionError();
         }
+        for (File file : fList)
+            if (file.getName().endsWith( ".txt" )) {
+                file.delete();
+
+            } else if (file.getName().startsWith( "control.txt" )) {
+
+            }
     }
 
 
     public static void grabWebPage(boolean useCustomURL) throws IOException {
         // Reads From Control File
-        clearFiles();
         String link = "";
         Document webDoc;
         Elements webElements;
@@ -56,101 +60,105 @@ public class HTMLParser {
 
 
         if (useCustomURL) {
-            /// Grab Text from gui
+            customLinkFlag = true;
+            /// Grab Text from GUI
             link = UI.getURLTxtBox();
 
-            webDoc = Jsoup.connect(link + "").userAgent("Mozilla").data("name", "jsoup").get();
-            webElements = webDoc.select("div#mw-content-text");
+            webDoc = Jsoup.connect( link + "" ).userAgent( "Mozilla" ).data( "name", "jsoup" ).get();
+            webElements = webDoc.select( "div#mw-content-text" );
 
             // To make File titles
-            String[] shortLink = link.split("^(.*[\\\\\\/])");
+            String[] shortLink = link.split( "^(.*[\\\\\\/])" );
 
             StringBuilder builder = new StringBuilder();
             for (String value : shortLink) {
-                builder.append(value);
+                builder.append( value );
             }
-            String text = builder.toString();
+            text = builder.toString();
             //This Makes the output  (Makes a output ctrlFile and cuts off System.out (GUI Depends on this being false))
-            PrintStream fileOut = new PrintStream(new File(text + ".txt"));
-           // System.out.println(text);
-
-            System.setOut(fileOut);
+            PrintStream fileOut = new PrintStream( new File( text + ".txt" ) );
+            System.setOut( fileOut );
             for (Element el : webElements) {
                 elements = el.text();
 
             }
-
-            //--DEBUG
-            //System.out.println(amountOfLinks);
-            // System.out.println(link);
-            printListVertically(filterAndSort(elements));
+            printListVertically( filterAndSort( elements ) );
 
 
-        } else if (!useCustomURL) {
-            for (amountOfLinks = 0; amountOfLinks < 10 && sc.hasNextLine(); amountOfLinks++) {
-               // System.out.println(amountOfLinks);
+        } else {
+            int controlFileLinkLimit = 100;
+            for (int amountOfLinks = 0; amountOfLinks < controlFileLinkLimit && scanControlFile.hasNextLine(); amountOfLinks++) {
+                link = scanControlFile.nextLine();
 
-                link = sc.nextLine();
 
-
-                webDoc = Jsoup.connect(link).userAgent("Mozilla").data("name", "jsoup").get();
-                webElements = webDoc.select("div#mw-content-text");
+                webDoc = Jsoup.connect( link ).userAgent( "Mozilla" ).data( "name", "jsoup" ).get();
+                webElements = webDoc.select( "div#mw-content-text" );
 
                 // To make File titles
-                String[] shortLink = link.split("^(.*[\\\\\\/])");
+                String[] shortLink = link.split( "^(.*[\\\\\\/])" );
 
                 StringBuilder builder = new StringBuilder();
                 for (String value : shortLink) {
-                    builder.append(value);
+                    builder.append( value );
                 }
-                String text = builder.toString();
+                text = builder.toString();
                 //This Makes the output  (Makes a output ctrlFile and cuts off System.out (GUI Depends on this being false))
-                PrintStream fileOut = new PrintStream(new File(text + ".txt"));
-                // System.out.println(text);
+                PrintStream fileOut = new PrintStream( new File( text + ".txt" ) );
 
-                System.setOut(fileOut);
+                System.setOut( fileOut );
                 for (Element el : webElements) {
                     elements = el.text();
 
                 }
-                //--DEBUG
-                //System.out.println(amountOfLinks);
-                // System.out.println(link);
-                printListVertically(filterAndSort(elements));
-                // System.out.println(link);
+                printListVertically( filterAndSort( elements ) );
+
             }
         }
     }
 
     // Parse and returns raw text
-    public static List<Map.Entry<String, Long>> filterAndSort(String string) {
+    private static List<Map.Entry<String, Long>> filterAndSort(String string) {
         //
-        Map<String, Long> frequencyMap = Arrays.stream(string.split("\\s+"))
-                .filter(word -> word.matches("\\b\\w{5,}\\b"))
-                .collect(Collectors.groupingBy(s -> s, Collectors.counting()));
-        Comparator<Map.Entry<String, Long>> byValue = Comparator.comparing(Map.Entry::getValue);
+        Map<String, Long> frequencyMap = Arrays.stream( string.split( "\\s+" ) )
+                .filter( word -> word.matches( "\\b\\w{5,}\\b" ) )
+                .collect( Collectors.groupingBy( s -> s, Collectors.counting() ) );
+        Comparator<Map.Entry<String, Long>> byValue = Comparator.comparing( Map.Entry::getValue );
         //
         List<Map.Entry<String, Long>> sortedByFrequency = frequencyMap.entrySet()
                 .stream()
-                .sorted(byValue.reversed())
-                .collect(Collectors.toList());
+                .sorted( byValue.reversed() )
+                .collect( Collectors.toList() );
         return sortedByFrequency;
     }
 
-    public static void printListVertically(List<Entry<String, Long>> listEntry) {
+    private static void printListVertically(List<Entry<String, Long>> listEntry) {
 
         for (Entry<String, Long> entry : listEntry) {
             String key = entry.getKey();
             Long value = entry.getValue();
-            System.out.println(key + " = " + value);
+            System.out.println( key + " = " + value );
         }
 
     }
 
-    // Constructor
-    public HTMLParser() throws IOException {
-        // UI supposed to set bool
+    public static void recommendPages() throws IOException {
 
+        if (customLinkFlag = true) {
+            Scanner fileReader = new Scanner( new File( text + ".txt" ) );
+            int significantWords = 5;
+            int words;
+            PrintStream fileOut = new PrintStream( new File( "Recommendations.txt" ) );
+            System.setOut( fileOut );
+            for (words = 0; fileReader.hasNextLine() == words < significantWords; words++) {
+                System.out.println( "https://en.wikipedia.org/wiki/" + fileReader.nextLine().replaceAll( "[^a-zA-Z]", "" ) );
+
+            }
+
+        }
+    }
+
+    // Constructor
+    public HTMLParser() {
     }
 }
